+++ lib/puree/extractor/publication.rb
@@ -16,14 +16,13 @@ module Puree
 
       def combine_metadata
         super
+
         @model.category = @extractor.category
         @model.description = @extractor.description
-        @model.doi = @extractor.doi
-        @model.event = @extractor.event
         @model.files = @extractor.files
         @model.keywords = @extractor.keywords
         @model.organisations = @extractor.organisations
-        @model.pages = @extractor.pages
+        @model.owner = @extractor.owner
         @model.persons_internal = @extractor.persons_internal
         @model.persons_external = @extractor.persons_external
         @model.persons_other = @extractor.persons_other