+++ lib/puree/model/publication.rb
@@ -12,10 +12,10 @@ module Puree
       attr_accessor :description
 
       # @return [String, nil]
-      attr_accessor :doi
+      # attr_accessor :doi
 
       # @return [Puree::Model::EventHeader, nil]
-      attr_accessor :event
+      # attr_accessor :event
 
       # @return [Array<Puree::Model::File>]
       attr_accessor :files
@@ -26,8 +26,11 @@ module Puree
       # @return [Array<Puree::Model::OrganisationHeader>]
       attr_accessor :organisations
 
+      # @return [Puree::Model::OrganisationHeader, nil]
+      attr_accessor :owner
+
       # @return [Fixnum, nil]
-      attr_accessor :pages
+      # attr_accessor :pages
 
       # @return [Array<Puree::Model::EndeavourPerson>]
       attr_accessor :persons_internal